+++ src/core/PathElevation.js
