+++ src/core/lang/en.json
