+++ src/core/utils/LogBuffer.js
