+++ src/core/utils/googleMapsClient.js
@@ -1,5 +1,5 @@
 import googleMaps from '@google/maps';
-import config from '../../server/config.json';
+import config from '../../../server/config.json';
 
 const googleMapsClient = googleMaps.createClient({
     key: config.googleMapsApiKey,
